docker image for gateway with mraa
This docker container image is for gateway with mraa
installation. It also has a led-blink program to test the
setup and instructions on how to run it.

The led-blink program is assuming Arduino 101 connected to
the gateway.

To keep the docker image size minimal almost all the dependency
installation happens when the image runs in a container. This
will take ~3 mins on a corei7 and ~10 mins on Atom gateway.

Signed-off-by: Huzefa N Kankroliwala <huzefank@gmail.com>